Implement basic diffusive lighting in volumetric raytracer
Approximated surface normals are encoded into the moments texture.
These normals are restored when the ray marcher encounters an
impermeable cell. Note that normals are shifted to be >= 0 as negative
values are not representable using OpenGL textures.

If this prototype works out it might be preferable to calculate better
surface normals during preprocessing to be stored separately.

Compare to 6123c8a.